+++ src/lightly_train/__init__.py
@@ -32,6 +32,7 @@ from lightly_train._commands.common_helpers import ModelFormat, ModelPart
 from lightly_train._commands.embed import embed
 from lightly_train._commands.export import export
 from lightly_train._commands.export_task import export_onnx
+from lightly_train._commands.predict_task import predict_semantic_segmentation
 from lightly_train._commands.train import train
 from lightly_train._commands.train_task import (
     train_object_detection,
@@ -54,6 +55,7 @@ __all__ = [
     "train",
     "train_semantic_segmentation",
     "train_object_detection",
+    "predict_semantic_segmentation",
     "load_model_from_checkpoint",
 ]