+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Object detection metrics tracking and required dependencies.
+- Add support for DINOv3 ConvNeXt models.
+- Automatically download DINOv3 weights.
+- Predict masks for a full dataset with a semantic segmentation model checkpoint with
+  the `predict_semantic_segmentation` command. It can be used for e.g. auto-labeling
+  datasets.
 
 ### Changed
 
@@ -21,6 +26,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+
+- Fix issue with loading DINOv3 SAT493m checkpoints.
+- Fixed an issue where dataset cache files were incorrectly saved.
 
 ## [0.11.4] - 2025-10-08